Examples/public-docker-example/requirements.txt
python-dotenv~=1.1
bithuman>=1.6.0
livekit-agents[openai,bithuman,silero]~=1.3
